fix preview to only erase one line for the spinner
Before this change, every 100ms we erase and reprint all preview
lines. This is very bad when the preview doesn't fit the screen, and
it's also problematic when ran via ssh.

We fix this by redesigning `ui.Writer`. It now provides the ability to
obtain "marks" and then clear back lines up to a specific mark. The
preview and the spinner use separate marks.